+++ hipparchus-core/src/main/java/org/hipparchus/linear/OpenMapRealVector.java
@@ -175,7 +175,7 @@ public class OpenMapRealVector extends SparseRealVector
         entries = new OpenIntToDoubleHashMap(0.0);
         this.epsilon = epsilon;
         for (int key = 0; key < values.length; key++) {
-            double value = values[key].doubleValue();
+            double value = values[key];
             if (!isDefaultValue(value)) {
                 entries.put(key, value);
             }
